+++ packages/multichain-account-service/src/logger.ts
@@ -1,0 +1,9 @@
+import { createProjectLogger, createModuleLogger } from '@metamask/utils';
+
+export const projectLogger = createProjectLogger('multichain-account-service');
+
+export { createModuleLogger };
+
+export const WARNING_PREFIX = 'WARNING --';
+
+export type Logger = typeof projectLogger;